requirements.txt
pandas==2.3.2
SQLAlchemy==2.0.43
python-dotenv==1.1.1
psycopg2-binary==2.9.10
pyodbc==5.2.0
